+++ CHANGES.rst
@@ -31,6 +31,9 @@ Imviz
 - Line profile plot in Line Profile plugin no longer affects
   radial profile plot in Simple Aperture Photometry plugin. [#1224]
 
+- Radial profile plot in Simple Aperture Photometry plugin
+  no longer shows masked aperture data. [#1224]
+
 Mosviz
 ^^^^^^